vehicle-signal-server-spec/0000-subscribe-standard-flow.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Vehicle Signal Server Spec Test: subscribe() test</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-helper.js"></script>

<h3> Subscribe Test: subscribe full process success</h3>
<h3> Subscribe Signal.Drivetrain.Transmission.Speed</h3>

<h3>
<div id="log"></div>
</h3>

<script>
var t = async_test("Subscribe test: subscribe full process success ");

var vehicle  = new WebSocket(VISS_URL, VISS_SUBPROTO);

var PH_START            = 0;
var PH_SENT_SUBSCRIBE   = 1;
var PH_SUBSCRIBED       = 2;
var PH_SENT_UNSUBSCRIBE = 3;
var PH_UNSUBSCRIBED     = 4;
var PH_END              = 5;

var NOTIFICATION_COUNT  = 3;

vehicle.onopen = function() {
  // request a signal
  var reqId = "reqid-12345";
  var path = "Signal.Drivetrain.Transmission.Speed";
  var notifCount = 0;
  var phase = PH_START;
  var subId = null;

  // process messages from the server
  vehicle.onmessage = t.step_func(
    function(event) {
      var msg = JSON.parse(event.data);

      if (phase == PH_START) {
        // nothing to do
      } else if (phase == PH_SENT_SUBSCRIBE) {
        // wait for subscribe response
        if (isSubscribeSuccessResponse(reqId, msg)) {
          addLogMessage("subscribeSuccessResponse received. subscribeId = " + msg.subscriptionId);
          subId = msg.subscriptionId
          phase = PH_SUBSCRIBED;
          //assert_true(true, "OK: "+path + ": value = " + msg.value);
        } else if (isSubscribeErrorResponse(reqId, msg)) {
          helper_terminate_failure("subscribeErrorResponse received. test failed.");
          phase = PH_END;
        } else {
          // ignore this case
        }

      } else if (phase == PH_SUBSCRIBED) {
        // wait subscriptionNotification

        if (isSubscriptionNotification(subId, msg)){
          // after received 3 notif, send unsubscribe
          if (notifCount > NOTIFICATION_COUNT) {
            // do unsubscribe
            addLogMessage("Send unsubscribe request");
            vehicle.send('{"action":"unsubscribe"'
                          +' ,"subscriptionId":"'+subId+'"'
                          +' ,"requestId":"'+reqId+'"'
                          +'}');
            phase = PH_SENT_UNSUBSCRIBE;
          }
          addLogMessage("subscriptionNotification received. value = " + msg.value);
          notifCount++;
        }

      } else if (phase == PH_SENT_UNSUBSCRIBE) {
        // wait for unsubscribe response
        if (isUnsubscribeSuccessResponse(reqId, subId, msg)) {
          // unsubscibe success
          addLogMessage("unsubscribeSuccessResponse received");
          phase = PH_UNSUBSCRIBED;
          addLogSuccess("subscribe/unsubscribe successed");
          t.step_timeout(function() {
            assert_true(true, "OK: "+path + ": subscriptionId = " + msg.subId);
            vehicle.close();
            t.done();
          }, TIME_FINISH_WAIT);

        } else if (isUnsubscribeErrorResponse(reqId, subId, msg)) {
          // unsubscribe failure
          helper_terminate_failure("unsubscribeErrorResponse received");
          // need unsubscribe() again?
          // ignore
        } else {
          // ignore this case
        }

      } else if (phase == PH_UNSUBSCRIBED) {
        // nothing to do
      }
    }
  );

  // in case server doesn't return response
  t.step_timeout(function() {
    if (phase < PH_SUBSCRIBED) {_
      helper_termintate_failure("Finish by timeout");
    }
  }, TIME_OUT_TIME);

  addLogMessage("<br>Send subscribe request: path="+path);
  vehicle.send('{"action":"subscribe","path":"'+ path +'" ,"requestId":"'+reqId+'"}');
  phase = PH_SENT_SUBSCRIBE;
}
</script>
